chore(mcp): move slab to new architecture
