+++ front/lib/actions/mcp_internal_actions/servers/index.ts
@@ -8,7 +8,6 @@ import { default as interactiveContentServer } from "@app/lib/actions/mcp_intern
 import { default as productboardServer } from "@app/lib/actions/mcp_internal_actions/servers/productboard";
 import { default as salesforceServer } from "@app/lib/actions/mcp_internal_actions/servers/salesforce";
 import { default as salesloftServer } from "@app/lib/actions/mcp_internal_actions/servers/salesloft";
-import { default as slabServer } from "@app/lib/actions/mcp_internal_actions/servers/slab";
 import { default as slideshowServer } from "@app/lib/actions/mcp_internal_actions/servers/slideshow";
 import { default as vantaServer } from "@app/lib/actions/mcp_internal_actions/servers/vanta";
 import type { AgentLoopContextType } from "@app/lib/actions/types";
@@ -60,6 +59,7 @@ import { default as sandboxServer } from "@app/lib/api/actions/servers/sandbox";
 import { default as schedulesManagementServer } from "@app/lib/api/actions/servers/schedules_management";
 import { default as searchServer } from "@app/lib/api/actions/servers/search";
 import { default as skillManagementServer } from "@app/lib/api/actions/servers/skill_management";
+import { default as slabServer } from "@app/lib/api/actions/servers/slab";
 import { default as slackBotServer } from "@app/lib/api/actions/servers/slack_bot";
 import { default as slackServer } from "@app/lib/api/actions/servers/slack_personal";
 import { default as snowflakeServer } from "@app/lib/api/actions/servers/snowflake";
