+++ front/lib/api/actions/servers/slab/utils.ts
@@ -1,4 +1,4 @@
-import type { SlabPost } from "./slab_types";
+import type { SlabPost } from "./types";
 
 export const ERROR_MESSAGES = {
   NO_ACCESS_TOKEN:
